second submit

+++ ft_strlen.c
@@ -6,7 +6,7 @@
 /*   By: ymori <ymori@student.42tokyo.jp>           +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/01/19 14:41:40 by ymori             #+#    #+#             */
-/*   Updated: 2021/01/20 01:10:02 by ymori            ###   ########.fr       */
+/*   Updated: 2021/02/03 21:18:46 by ymori            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,10 +14,10 @@
 
 size_t	ft_strlen(const char *s)
 {
-	int		i;
+	size_t	i;
 
 	i = 0;
 	while (s[i] != '\0')
 		i++;
-	return ((size_t)i);
+	return (i);
 }
